Divide Big Cells
Divide Big Cells 
This script will check for cells that are “too large” to be displayed (or printed?). 
It is non-destructive, so you can run it at will. 
It checks every cell for text that is greater than 1000 bytes in total. It will divide those cells into chunks of  approximately 1000 bytes, and write these to a text file. It will then open Notepad. At this point, you can cut  and paste the sections of text into the source spreadsheet. All lines are written in their entirety (no lines are  broken up). 
Features: 
∙ Create a text file 
∙ Open the output text file in Notepad